refactor(platform): move file object identity comparison to platform/fs
